nanokexec: Use boot_param->init_size for kernel kexec memory size
The current implementation uses the ZO_INIT_SIZE for kernel memory
size  which is wrong, the correct way is to use INIT_SIZE which is
max(ZO_INIT_SIZE, VO_INIT_SIZE), and align it with kernel_alignment
to avoid taint the initramfs put behind.

So we modify the generate_kexec_segment to take care of two kinds of
size: "buf_len" and "mem_len".

This patch also added some debug print for the kexec_load syscall.

Signed-off-by: Xunlei Pang <xlpang@redhat.com>